Extracting out Motivation Factory
