+++ src/main/java/zd/zero/waifu/motivator/plugin/motivation/MotivationLifecycleListener.kt
@@ -1,0 +1,14 @@
+package zd.zero.waifu.motivator.plugin.motivation
+
+val defaultListener = object : MotivationLifecycleListener {
+    override fun onDisplay() {}
+
+    override fun onDispose() {}
+}
+
+interface MotivationLifecycleListener {
+
+    fun onDisplay()
+
+    fun onDispose()
+}
